+++ frontend/catalyst/jit.py
@@ -244,7 +244,7 @@ def qjit(
 
 
     .. details::
-        :title: In-place JAX array assignments with Autograph
+        :title: In-place JAX array updates with Autograph
 
         To update array values when using JAX, the JAX syntax for array assignment
         (which uses the array ``at`` and ``set`` methods) must be used:
@@ -278,6 +278,41 @@ def qjit(
 
         Under the hood, Catalyst converts anything coming in the latter notation into the
         former one.
+
+        Similarly, to update array values with an operation when using JAX, the JAX syntax for array
+        update (which uses the array ``at`` and the ``add``, ``sub``, etc. methods) must be used:
+
+        .. code-block:: python
+
+            @qjit(autograph=True)
+            def f(x):
+                first_dim = x.shape[0]
+                result = jnp.copy(x)
+
+                for i in range(first_dim):
+                    result = result.at[i].multiply(2)
+
+                return result
+
+        Again, if updating a single index of the array, Autograph supports conversion of
+        standard Python array operator assignment syntax for the equivalent in-place expressions
+        listed in the JAX documentation for ``jax.numpy.ndarray.at``:
+
+        .. code-block:: python
+
+            @qjit(autograph=True)
+            def f(x):
+                first_dim = x.shape[0]
+                result = jnp.copy(x)
+
+                for i in range(first_dim):
+                    result[i] *= 2
+
+                return result
+
+        Under the hood, Catalyst converts anything coming in the latter notation into the
+        former one.
+
 
 
     .. details::