+++ doc/changelog.md
@@ -81,6 +81,30 @@
 
   ```
 
+* Support for usage of single index JAX array operator update
+  inside Autograph annotated functions.
+  [(#769)](https://github.com/PennyLaneAI/catalyst/pull/769)
+
+  Using operator assignment syntax in favor of at...operation expressions is now possible for the following operations:
+  * `x[i] += y` in favor of `x.at[i].add(y)`
+  * `x[i] -= y` in favor of `x.at[i].add(-y)`
+  * `x[i] *= y` in favor of `x.at[i].multiply(y)`
+  * `x[i] /= y` in favor of `x.at[i].divide(y)`
+  * `x[i] **= y` in favor of `x.at[i].power(y)`
+
+  ```py
+  @qjit(autograph=True)
+  def f(x):
+    first_dim = x.shape[0]
+    result = jnp.copy(x)
+
+    for i in range(first_dim):
+      result[i] *= 2  # This is now supported
+
+    return result
+
+  ```
+
 <h3>Improvements</h3>
 
 * Catalyst now has support for `qml.sample(m)` where `m` is the result of a mid-circuit